Fix up debian/shims/chromium.sh
We can get rid of repeated calls to the external expr
command by using a case statement to match the files
that we want to avoid sourcing.

We can exit early if running without gdb in debugging
mode, and thereby avoid executing 'mktemp -d'. Note
that doing this avoids leaving behind a stray
temporary profile directory.

We can simplify the use of the exec command to start
chromium by taking the non-exec version out of the
else branch of the 'if..then' test (nothing below
the exec statement will get executed).

We can also factor out the 'rm -rf $TEMP_PROFILE'
command and execute it just before the end if
want_temp_profile=1.

the 'exit $?' command needs to be moved to the end of
the script so that both branches of the want_debug
test can get a proper exit code.

We also make the spacing more consistent, replace
backticks with '$()', and where appropriate convert
some 'if..then' commands to case statements.

There should be practically no functional change.

+++ debian/shims/chromium.sh
@@ -48,40 +48,46 @@ The hardware on this system lacks support for the sse3 instruction set.
 The upstream chromium project no longer supports this configuration.
 For more information, please go to https://crbug.com/1123353."
 
-case `uname -m` in
-    i386|i586|i686|x86_64)
-        # Check whether this system supports sse3
-        if ! grep -q sse3 /proc/cpuinfo; then
-            xmessage "$nosse3"
-            exit 1
-        fi
-        ;;
+case $(uname -m) in
+  i386|i586|i686|x86_64)
+    # Check whether this system supports sse3
+    if ! grep -q sse3 /proc/cpuinfo; then
+      xmessage "$nosse3"
+      exit 1
+    fi ;;
 esac
 
 # Source additional settings
 for file in /etc/chromium.d/*; do
-  test $file = /etc/chromium.d/README || expr $file : .*\.dpkg > /dev/null || . $file
+  case $file in
+    /etc/chromium.d/*.dpkg-*|/etc/chromium.d/README)
+      continue ;;
+  esac
+  . $file
 done
 
 # Use the /usr/bin helper script for generated launchers
-if test -z "$CHROME_WRAPPER"; then
-    export CHROME_WRAPPER="/usr/bin/$APPNAME"
-fi
+case "$CHROME_WRAPPER" in
+  "")
+    export CHROME_WRAPPER="/usr/bin/$APPNAME" ;;
+esac
 
 # Set the correct file name for the desktop file
 export CHROME_DESKTOP="chromium.desktop"
 
 # Set CHROME_VERSION_EXTRA text, which is displayed in the About dialog
-DIST=`print_dist`
+DIST=$(print_dist)
 BUILD_DIST="@BUILD_DIST@"
 export CHROME_VERSION_EXTRA="built on $BUILD_DIST, running on $DIST"
 
 # Add LIBDIR to LD_LIBRARY_PATH to load libffmpeg.so (if built as a component)
-if [ -z "${LD_LIBRARY_PATH:+nonempty}" ] ; then
-    LD_LIBRARY_PATH=$LIBDIR
-else
-    LD_LIBRARY_PATH=$LIBDIR:$LD_LIBRARY_PATH
-fi
+case "${LD_LIBRARY_PATH:+nonempty}" in
+  "")
+    LD_LIBRARY_PATH=$LIBDIR ;;
+
+  *)
+    LD_LIBRARY_PATH=$LIBDIR:$LD_LIBRARY_PATH ;;
+esac
 
 export LD_LIBRARY_PATH
 
@@ -109,18 +115,19 @@ while [ $# -gt 0 ]; do
   esac
 done
 
-if [ $want_temp_profile -eq 1 ] ; then
-  TEMP_PROFILE=`mktemp -d`
-  echo "Using temporary profile: $TEMP_PROFILE"
-  CHROMIUM_FLAGS="$CHROMIUM_FLAGS --user-data-dir=$TEMP_PROFILE"
+if [ $want_debug -eq 1 ] && [ ! -x $GDB ]; then
+  echo "Sorry, can't find usable $GDB. Please install it."
+  exit 1
 fi
 
-if [ $want_debug -eq 1 ] ; then
-  if [ ! -x $GDB ] ; then
-    echo "Sorry, can't find usable $GDB. Please install it."
-    exit 1
-  fi
-  tmpfile=`mktemp /tmp/chromiumargs.XXXXXX` || { echo "Cannot create temporary file" >&2; exit 1; }
+if [ $want_temp_profile -eq 1 ]; then
+  TEMP_PROFILE=$(mktemp -d)
+  CHROMIUM_FLAGS="$CHROMIUM_FLAGS --user-data-dir=$TEMP_PROFILE"
+  echo "Using temporary profile: $TEMP_PROFILE"
+fi
+
+if [ $want_debug -eq 1 ]; then
+  tmpfile=$(mktemp /tmp/chromiumargs.XXXXXX) || { echo "Cannot create temporary file" >&2; exit 1; }
   trap " [ -f \"$tmpfile\" ] && /bin/rm -f -- \"$tmpfile\"" 0 1 2 3 13 15
   echo "set args $CHROMIUM_FLAGS --single-process ${1+"$@"}" > $tmpfile
   echo "# Env:"
@@ -130,16 +137,12 @@ if [ $want_debug -eq 1 ] ; then
   echo "#      CHROMIUM_FLAGS=$CHROMIUM_FLAGS"
   echo "$GDB $LIBDIR/$BINNAME -x $tmpfile"
   $GDB "$LIBDIR/$BINNAME" -x $tmpfile
-  if [ $want_temp_profile -eq 1 ] ; then
-    rm -rf $TEMP_PROFILE
-  fi
-  exit $?
 else
-  if [ $want_temp_profile -eq 0 ] ; then
-    exec $LIBDIR/$BINNAME $CHROMIUM_FLAGS "$@"
-  else
-    # we can't exec here as we need to clean-up the temporary profile
-    $LIBDIR/$BINNAME $CHROMIUM_FLAGS "$@"
-    rm -rf $TEMP_PROFILE
-  fi
+  # Use exec here as we will have no $TEMP_PROFILE to later delete
+  [ $want_temp_profile -eq 1 ] || exec $LIBDIR/$BINNAME $CHROMIUM_FLAGS "$@"
+  $LIBDIR/$BINNAME $CHROMIUM_FLAGS "$@"
 fi
+
+[ $want_temp_profile -eq 0 ] || rm -rf $TEMP_PROFILE
+
+exit $?
